+++ packages/eas-cli/src/commandUtils/EasCommand.ts
@@ -21,13 +21,16 @@ export default abstract class EasCommand extends Command {
    * force the user to log in
    */
   protected requiresAuthentication = true;
+  protected mustBeRunInsideProject = true;
 
   protected abstract runAsync(): Promise<any>;
 
   // eslint-disable-next-line async-protect/async-suffix
   async run(): Promise<any> {
     await initAnalyticsAsync();
-    await this.applyCliConfigAsync();
+    if (this.mustBeRunInsideProject) {
+      await this.applyCliConfigAsync();
+    }
 
     if (this.requiresAuthentication) {
       const { flags } = this.parse();